Move AdjointFixedAutograd to a separate file
